Flanger: the wet interpolation needs 64 bits, and a fixture that says so
Review finding on the port, not something the goldens could have caught: they
matched to the byte.

`s0 + ((s1 - s0) * delay_frac >> 16)` was computed in `int32_t`. The CPython
twin computes it in Python, which does not overflow. With the delay line
holding adjacent samples at opposite rails `s1 - s0` reaches 65535, and
`delay_frac` reaches 65535 too - a product of 4.29e9, past INT32_MAX. That is
signed overflow, and it is reachable: a full-scale alternating source diverged
from the twin on every setting tried.

    source: +/-32767 alternating, 8 kHz mono, three settings

                     CPython        MicroPython (int32)   MicroPython (int64)
    defaults         58306b0c       ab46fe89              58306b0c
    deep+fast        9003899b       76e0e18d              9003899b
    full feedback    d129d107       ee9d263d              d129d107

`granular_pitch_shift` was checked for the same shape and does not have it: its
read shift is 8, so `frac` tops out at 255 and the product stays inside int32
by three orders of magnitude, and its grain interpolation is a convex
combination that cannot leave int16 range before the envelope multiply.

Three `rails` cases added to `flanger_probe.py`, and they are red on the
unwidened code - verified by putting `int32_t` back, rebuilding, and reading
`FAIL flanger_probe.py micropython got aa1060ee... expected 824037d7...`, then
restoring and rebuilding to `ok` on both interpreters. Every pre-existing line
of that probe's output is byte-identical before and after; one golden key moves,
the other nineteen untouched.

Worth saying plainly, because it is the argument for cross-interpreter checking
over stored digests: the stored golden matched perfectly while the two
interpreters disagreed. Only comparing the targets against each other found it.

+++ src/shared/audioif_flanger.c
@@ -51,7 +51,15 @@ void audioif_flanger_process_s16(int16_t *output, const int16_t *input,
         uint32_t r1 = (r0 == 0) ? frames - 1 : r0 - 1;
         int32_t s0 = delay_line[plane + r0];
         int32_t s1 = delay_line[plane + r1];
-        int32_t wet = s0 + (((s1 - s0) * (int32_t)delay_frac) >> 16);
+        // 64-bit for the same reason the span/tri product above is: the
+        // twin computes this in Python, which does not overflow. `s1 - s0`
+        // reaches +/-65535 when adjacent line samples sit at opposite rails,
+        // and 65535 * 65535 is 4.29e9 - past INT32_MAX. In int32 that is
+        // signed overflow, and it is reachable on ordinary material: a
+        // full-scale alternating source diverged from the CPython twin on
+        // every setting tried until this cast.
+        int32_t wet = s0 +
+            (int32_t)((((int64_t)s1 - (int64_t)s0) * (int64_t)delay_frac) >> 16);
         delay_line[plane + write] = audioif_sat16(
             sample + audioif_sat16(wet * feedback, 15), 0);
         state->write_pos[channel] = (write + 1 == frames) ? 0 : write + 1;
